move ProblemList, document, and fix Guide module docstring
