+++ analysis/monkbrewmaster/src/Guide.tsx
@@ -2,7 +2,7 @@ import colorForPerformance from 'common/colorForPerformance';
 import SPELLS from 'common/SPELLS';
 import { SpellLink } from 'interface';
 import { GuideProps, PassFailBar, Section, SubSection } from 'interface/guide';
-import ProblemList, { Problem, ProblemRendererProps } from 'interface/guide/Problems';
+import ProblemList, { Problem, ProblemRendererProps } from 'interface/guide/ProblemList';
 import { AnyEvent, DamageEvent } from 'parser/core/Events';
 import { Info } from 'parser/core/metric';
 import CastEfficiency from 'parser/shared/modules/CastEfficiency';
